turret and enemy spawn logic
turret will now face the right direction of the enemy when it gets in
range.
enemy will now spawn at grid location (3,0) every 2 seconds and run to
the target location (i set a target location by default).

-fixed a bug where if the actor is on the edge of the window and a wall
is right across from it the program will crash(eg: player on left side
of window, and a wall is directly to the right of the player).